+++ src/player_interaction/building/components.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::*;
+
+/// Component to attach to any bulding
+#[derive(Component)]
+pub struct Building {}
